Significant rewrite of ldap_servers and dependencies.
This commit removes always requiring the procedural helper functions and moves
them into separate classes. The basic LDAP return values are now encapsulated
in an interface. This means that the majority of global constants are gone.
Additionally an erroneous copy of the LdapTypes from the previous commit is
removed as well. Please note that several unused functions were removed.