Major update to global variables and test data
This commit removes all major global constants so that autoloading is fully
possible for the test runner, which also makes the code significantly easier to
read. This also required wrapping the loading of the server connection in a
separate class (ServerFactory).

This commit also removes the ldap_test module since it present a secondary
approach to manual testing which is not within useful scope of the module. The
base test classes were moved to ldap_servers. The documentation needs to bei
updated for this.

State of the testing infrastructure: The test server can in theory be used
but the gap between the inferred structure and current structure of the system
configuration presents a gap that is not quickly bridged. The currently passing
tests only return without assertions.

Also some minor updates to ldap_query, still a work in progress.