Issue #2823609 by webflo, grahl: Provisioning new user broken via refactoring
This commit removes LdapUserConfAdmin and moves it primarily into the
administration form, for which it was a wrapper and the rest moves to
LdapUserConf where necessary. Significant reductions in complexity could be
reached in LdapUserConf by relying on Drupal::config, this process is not
complete. All major issues with the ldap_user admin form seem to have been
resolved as well.

Furthermore, the orphan functions were collated to allow for